// src/app/countries/interfaces/Cache-store.interface.ts
import { Country } from './Country.interface';
import { Region } from './Region.type';

export interface CacheStore {
  byCapital: TermCountries;
  byCountries: TermCountries;
  byRegion: RegionCountries;
}

export interface TermCountries {
  term: string;
  countries: Country[];
}

export interface RegionCountries {
  region: Region;
  countries: Country[];
}